+++ Tests/SentryObjCTests/SentryObjCFeedbackApiTests.m
@@ -40,68 +40,6 @@
                  }];
 }
 
-- (void)testFormViewController_shouldReturnUIViewController
-{
-    // -- Act --
-    UIViewController *viewController = [SentryObjCSDK.feedback formViewController];
-
-    // -- Assert --
-    XCTAssertNotNil(viewController);
-    XCTAssertTrue([viewController isKindOfClass:UIViewController.class]);
-}
-
-- (void)testFormViewControllerWithScreenshot_shouldReturnUIViewController
-{
-    // -- Arrange --
-    UIImage *screenshot = [[UIImage alloc] init];
-
-    // -- Act --
-    UIViewController *viewController =
-        [SentryObjCSDK.feedback formViewControllerWithScreenshot:screenshot];
-
-    // -- Assert --
-    XCTAssertNotNil(viewController);
-    XCTAssertTrue([viewController isKindOfClass:UIViewController.class]);
-}
-
-- (void)testFormViewControllerWithConfigure_shouldCallConfigurationCallback
-{
-    // -- Arrange --
-    __block BOOL configureCalled = NO;
-
-    // -- Act --
-    UIViewController *viewController = [SentryObjCSDK.feedback
-        formViewControllerWithConfigure:^(SentryObjCUserFeedbackConfiguration *configuration) {
-            configureCalled = YES;
-            configuration.animations = NO;
-        }];
-
-    // -- Assert --
-    XCTAssertTrue(configureCalled);
-    XCTAssertNotNil(viewController);
-    XCTAssertTrue([viewController isKindOfClass:UIViewController.class]);
-}
-
-- (void)testFormViewControllerWithScreenshotAndConfigure_shouldCallConfigurationCallback
-{
-    // -- Arrange --
-    UIImage *screenshot = [[UIImage alloc] init];
-    __block BOOL configureCalled = NO;
-
-    // -- Act --
-    UIViewController *viewController = [SentryObjCSDK.feedback
-        formViewControllerWithScreenshot:screenshot
-                               configure:^(SentryObjCUserFeedbackConfiguration *configuration) {
-                                   configureCalled = YES;
-                                   configuration.animations = NO;
-                               }];
-
-    // -- Assert --
-    XCTAssertTrue(configureCalled);
-    XCTAssertNotNil(viewController);
-    XCTAssertTrue([viewController isKindOfClass:UIViewController.class]);
-}
-
 @end
 
 #endif